+++ src/main/ruby/truffleruby/core/dir.rb
@@ -284,19 +284,23 @@ class Dir
 
     def chdir(path = ENV['HOME'])
       path = Truffle::Type.coerce_to_path path
+      path = path.dup.force_encoding(Encoding::LOCALE) if path.encoding == Encoding::BINARY
 
       if block_given?
         original_path = self.getwd
+        TrufflePrimitive.dir_set_truffle_working_directory(path)
         ret = Truffle::POSIX.chdir path
         Errno.handle(path) if ret != 0
 
         begin
           yield path
         ensure
+          TrufflePrimitive.dir_set_truffle_working_directory(original_path)
           ret = Truffle::POSIX.chdir original_path
           Errno.handle(original_path) if ret != 0
         end
       else
+        TrufflePrimitive.dir_set_truffle_working_directory(path)
         ret = Truffle::POSIX.chdir path
         Errno.handle path if ret != 0
         ret